move release.batch.io classes to subfolders
