+++ eva-accession-release/src/main/java/uk/ac/ebi/eva/accession/release/batch/io/merged/MergedVariantContextWriter.java
@@ -13,16 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.release.batch.io;
+package uk.ac.ebi.eva.accession.release.batch.io.merged;
 
 import htsjdk.variant.vcf.VCFHeaderLine;
 import htsjdk.variant.vcf.VCFHeaderLineType;
 import htsjdk.variant.vcf.VCFInfoHeaderLine;
 
+import uk.ac.ebi.eva.accession.release.batch.io.active.VariantContextWriter;
+import uk.ac.ebi.eva.accession.release.batch.io.VariantMongoAggregationReader;
+
 import java.nio.file.Path;
 import java.util.Set;
-
-import static uk.ac.ebi.eva.accession.release.batch.io.MergedVariantMongoReader.MERGED_INTO_KEY;
 
 public class MergedVariantContextWriter extends VariantContextWriter {
 
@@ -33,7 +34,7 @@ public class MergedVariantContextWriter extends VariantContextWriter {
     @Override
     protected Set<VCFHeaderLine> buildHeaderLines() {
         Set<VCFHeaderLine> vcfHeaderLines = super.buildHeaderLines();
-        vcfHeaderLines.add(new VCFInfoHeaderLine(MERGED_INTO_KEY, 1, VCFHeaderLineType.String,
+        vcfHeaderLines.add(new VCFInfoHeaderLine(VariantMongoAggregationReader.MERGED_INTO_KEY, 1, VCFHeaderLineType.String,
                                                  "RS ID that currently represents the variant"));
         return vcfHeaderLines;
     }
