+++ eva-accession-release/src/main/java/uk/ac/ebi/eva/accession/release/batch/io/multimap/MultimapVariantContextWriter.java
@@ -13,11 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.release.batch.io;
+package uk.ac.ebi.eva.accession.release.batch.io.multimap;
 
 import htsjdk.variant.vcf.VCFHeaderLine;
 import htsjdk.variant.vcf.VCFHeaderLineType;
 import htsjdk.variant.vcf.VCFInfoHeaderLine;
+
+import uk.ac.ebi.eva.accession.release.batch.io.active.VariantContextWriter;
 
 import java.nio.file.Path;
 import java.util.Set;
