+++ eva-accession-release/src/main/java/uk/ac/ebi/eva/accession/release/configuration/batch/io/MergedVariantMongoReaderConfiguration.java
@@ -27,7 +27,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
 import uk.ac.ebi.eva.accession.core.configuration.nonhuman.MongoConfiguration;
-import uk.ac.ebi.eva.accession.release.batch.io.MergedVariantMongoReader;
+import uk.ac.ebi.eva.accession.release.batch.io.merged.MergedVariantMongoReader;
 import uk.ac.ebi.eva.accession.release.parameters.InputParameters;
 import uk.ac.ebi.eva.commons.batch.io.UnwindingItemStreamReader;
 import uk.ac.ebi.eva.commons.core.models.pipeline.Variant;
