+++ eva-accession-release/src/test/java/uk/ac/ebi/eva/accession/release/batch/io/contig/ContigWriterTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.release.batch.io;
+package uk.ac.ebi.eva.accession.release.batch.io.contig;
 
 import org.junit.Before;
 import org.junit.Rule;
@@ -22,6 +22,7 @@ import org.junit.rules.TemporaryFolder;
 
 import uk.ac.ebi.eva.accession.core.contig.ContigMapping;
 import uk.ac.ebi.eva.accession.core.contig.ContigSynonyms;
+import uk.ac.ebi.eva.accession.release.batch.io.contig.ContigWriter;
 
 import java.io.BufferedReader;
 import java.io.File;
