+++ eva-accession-release/src/test/java/uk/ac/ebi/eva/accession/release/configuration/batch/steps/ListMultimapContigsStepConfigurationTest.java
@@ -48,7 +48,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import static org.junit.Assert.assertEquals;
-import static uk.ac.ebi.eva.accession.release.batch.io.ContigWriter.getMultimapContigsFilePath;
+import static uk.ac.ebi.eva.accession.release.batch.io.contig.ContigWriter.getMultimapContigsFilePath;
 import static uk.ac.ebi.eva.accession.release.configuration.BeanNames.LIST_MULTIMAP_CONTIGS_STEP;
 
 @RunWith(SpringRunner.class)
